Basic solution
This commit showcases a basic solution to this exercise using only
Spring Framework's test utilities.

MockRestServiceServer can be configured to define a set of expectations
based on a RestTemplate instance. When the client uses the RestTemplate
to issue requests to external services, the mock records the
interactions, validate they match the expectations and return the
configured result.

Because the RestTemplate instance has to be shared between the two
components, an extra package protected constructor had to be created.