Solution
MockRestServiceServer can be configured to define a set of expectations
based on a RestTemplate instance. When the client uses the RestTemplate
to issue requests to external services, the mock records the
interactions, validate they match the expectations and return the
configured result.

This commit uses Spring Boot's testing facilities. Contrary to a basic
solution with Spring Framework alone, this commit greatly simplifies the
setup of the MockRestServiceServer.

@RestClientTest is a dedicated test annotation for rest client. When it
is used on a given class, the client is managed by the context and
automatically configured to use a matching MockRestServiceServer.

All that is required is to inject the components in the test to use
them.